Add car component
